requirements.txt
# Core
streamlit>=1.29.0
python-dotenv>=1.0.0

# Data processing and visualization
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0
altair>=5.2.0

# NLP and ML
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
textstat>=0.7.0

# HTTP requests
requests>=2.31.0

# Database
psycopg2-binary>=2.9.0
